Pathology reworking
