+++ data_processing/pathology/cli/examples/collect_tiles.json
@@ -1,0 +1,5 @@
+{
+  "input_label_tag": "test_save_tiles",
+  "job_tag": "test_collect_tiles",
+  "output_container": "test_output_dataset"
+}
