+++ data_processing/pathology/cli/examples/generate_tile_labels.json
