+++ data_processing/pathology/cli/examples/save_tiles.json
@@ -1,6 +1,5 @@
 {
   "input_wsi_tag": "pathology.etl", 
   "input_label_tag": "test_generate_tiles",
-  "job_tag": "test_collect_tiles",
-  "output_container": "test_output_dataset"
+  "job_tag": "test_save_tiles"
 }
