+++ data_processing/pathology/cli/examples/visualize_tile_labels.json
