+++ data_processing/pathology/cli/save_tiles.py
@@ -3,21 +3,15 @@ Created: February 2021
 @author: aukermaa@mskcc.org
 
 Given a slide (container) ID
-1. resolve the path to the WSI image
+1. resolve the path to the WsiImage and TileLabels
 2. perform various scoring and labeling to tiles
-3. save tiles as a csv with schema [address, coordinates, *scores, *labels ]
+3. save tiles as a parquet file with schema [address, coordinates, *scores, *labels ]
 
 Example:
-python3 -m data_processing.pathology.cli.generate_tile_labels \
+python3 -m data_processing.pathology.cli.collect_tiles \
     -c TCGA-BRCA \
     -s tcga-gm-a2db-01z-00-dx1.9ee36aa6-2594-44c7-b05c-91a0aec7e511 \
-    -m data_processing/pathology/cli/example_generate_tile_labels.json
-
-Example with annotation:
-python3 -m data_processing.pathology.cli.generate_tile_labels \
-        -c ov-path-druv  \
-        -s 226871 \
-        -m data_processing/pathology/cli/example_generate_tile_labels.json 
+    -m data_processing/pathology/cli/example_collect_tiles.json
 '''
 
 # General imports
@@ -31,10 +25,9 @@ from data_processing.common.Container       import Container
 from data_processing.common.Node            import Node
 from data_processing.common.config          import ConfigSet
 
-# From radiology.common
-from data_processing.pathology.common.preprocess   import pretile_scoring
+from data_processing.pathology.common.preprocess import save_tiles
 
-logger = init_logger("generate_tile_labels.log")
+logger = init_logger("save_tiles.log")
 cfg = ConfigSet("APP_CFG",  config_file="config.yaml")
 
 @click.command()
@@ -44,22 +37,23 @@ cfg = ConfigSet("APP_CFG",  config_file="config.yaml")
 def cli(cohort_id, container_id, method_param_path):
     with open(method_param_path) as json_file:
         method_data = json.load(json_file)
-    generate_tile_labels_with_container(cohort_id, container_id, method_data)
+    save_tiles_with_container(cohort_id, container_id, method_data)
 
-def generate_tile_labels_with_container(cohort_id: str, container_id: str, method_data: dict):
+def save_tiles_with_container(cohort_id: str, container_id: str, method_data: dict):
     """
-    Using the container API interface, score and generate tile addresses
+    Using the container API interface, visualize tile-wise scores
     """
 
     # Do some setup
-    container   = Container( cfg ).setNamespace(cohort_id).lookupAndAttach(container_id)
-    method_id   = method_data.get("job_tag", "none")
+    container = Container( cfg ).setNamespace(cohort_id).lookupAndAttach(container_id)
+
+    method_id  = method_data.get("job_tag", "none")
 
     image_node  = container.get("WholeSlideImage", method_data['input_wsi_tag']) 
-    
-    # get image_id
-    # TODO - allow -s to take in slide (container) id
-    image_id = container._name
+    label_node  = container.get("TileScores", method_data['input_label_tag']) 
+
+    # Add properties to method_data
+    method_data.update(label_node.properties)
 
     try:
         if image_node is None:
@@ -70,12 +64,13 @@ def generate_tile_labels_with_container(cohort_id: str, container_id: str, metho
         output_dir = os.path.join(os.environ['MIND_GPFS_DIR'], "data", container._namespace_id, container._name, method_id)
         if not os.path.exists(output_dir): os.makedirs(output_dir)
 
-        properties = pretile_scoring(image_node.data, output_dir, method_data, image_id)
+        properties = save_tiles(image_node.data, label_node.data, output_dir, method_data )
 
     except Exception:
         container.logger.exception ("Exception raised, stopping job execution.")
     else:
-        output_node = Node("TileScores", method_id, properties)
+
+        output_node = Node("TileImages", method_id, properties)
         container.add(output_node)
         container.saveAll()
