+++ data_processing/pathology/cli/visualize_tile_labels.py
@@ -56,10 +56,9 @@ def visualize_tile_labels_with_container(cohort_id: str, container_id: str, meth
     method_id   = method_data.get("job_tag", "none")
     
     image_node  = container.get("WholeSlideImage", method_data['input_wsi_tag']) 
-    label_node  = container.get("TileScores", method_data['input_label_tag']) 
+    label_node  = container.get("TileScores",      method_data['input_label_tag']) 
 
     method_data.update(label_node.properties)
-
 
     try:
         if image_node is None:
@@ -70,7 +69,7 @@ def visualize_tile_labels_with_container(cohort_id: str, container_id: str, meth
         output_dir = os.path.join(os.environ['MIND_GPFS_DIR'], "data", container._namespace_id, container._name, method_id)
         if not os.path.exists(output_dir): os.makedirs(output_dir)
 
-        properties = visualize_scoring(image_node.get_path(), label_node.get_path(), output_dir, method_data)
+        properties = visualize_scoring(image_node.data, label_node.data, output_dir, method_data)
 
     except Exception:
         container.logger.exception ("Exception raised, stopping job execution.")
